feat(components): add user and therapist component, add profie page
user role #90
